Made updates based on udacity feedback from project submission
